dgo-gen: use uint32 to store function id

+++ dgo-gen/internal/collector/context.go
@@ -6,7 +6,7 @@ import (
 
 type _IdManager struct {
 	nTypes   int64
-	nMethods int64
+	nMethods uint32
 }
 
 func (m *_IdManager) NextTypeId() int64 {
@@ -14,7 +14,7 @@ func (m *_IdManager) NextTypeId() int64 {
 	return m.nTypes
 }
 
-func (m *_IdManager) NextMethodId() int64 {
+func (m *_IdManager) NextMethodId() uint32 {
 	m.nMethods += 1
 	return m.nMethods
 }
